android: use new environment.conf file to read in required env vars
The Android app now provides a file called environment.conf which includes
the env vars that pinentry needs in order to find all the .so and .jar
files that are needed to run.  These env vars are unset as part of the
env sanitizing, or are not provided by the Android sytem in the form that
they are needed.